Create Search Recyclables
This is the .html file for the creation of the search bar. What is in the comments is there to be tested functionality later on, once the main page is fully working. I am still doing React tutorials and will look to see how I can implement it into the product. I have created the SQL database, however, as I now have access to the ecowebhosting I mentioned, I will have access to phpmyadmin which will make it a lot easier to retrieve the data from the database. I will see how I can post the database stuff in here too; will probably just be the queries themselves.