Create CSS -Search Recyclables
This is the CSS file linked to the .html file added to my branch [Jessica's branch]. Again this is a work-in-progress and will be added to as necessary. So far, if both files are ran, the search bar functionality appears. But will not yet work as there is no .php yet to connect to the database.